services/chunker/requirements.txt
grpcio==1.48.2
grpcio-tools==1.48.2
protobuf>=3.19.0,<3.20
